Feat: I just tidy up i lttle bit the code because its was very messy
